Add DSL for commands to define options
This changed many commands because it changed the signature of the run
method.

This exposed a hole in testing these where there are no tests that
confirm the run method has a signature that matches what the command
runner will pass in. I'm considering fixing this in another commit by
making command tests always call through a runner.

Fixes #31.